Move ServiceDetails to the model package
This eliminates a package dependency cycle between the base
"dropwizard.error" package and the "dropwizard.error.health" package.

* Move ServiceDetails and ServiceDetailsTest to model package
* Update imports
* Add @AllArgsConstructor to ServiceDetails since previously it was
  package-private but is used by several tests. Decided for this simple
  a class it is fine to have the all-args constructor public, as I
  doubt it is going to change much if ever.
* Suppress empty method warnings (since they are intentionally empty)
  in ApplicationErrorExtensionUnitTest

Fixes #35
