+++ src/test/java/org/kiwiproject/dropwizard/error/Jdbi3ErrorContextTest.java
@@ -21,6 +21,7 @@ import org.kiwiproject.dropwizard.error.dao.jdbi3.Jdbi3ApplicationErrorDao;
 import org.kiwiproject.dropwizard.error.health.RecentErrorsHealthCheck;
 import org.kiwiproject.dropwizard.error.model.ApplicationError;
 import org.kiwiproject.dropwizard.error.model.DataStoreType;
+import org.kiwiproject.dropwizard.error.model.ServiceDetails;
 import org.kiwiproject.dropwizard.error.resource.ApplicationErrorResource;
 import org.kiwiproject.dropwizard.error.resource.GotErrorsResource;
 import org.kiwiproject.test.dropwizard.mockito.DropwizardMockitoMocks;
